+++ nestjs-BE/server/src/auth/auth.controller.spec.ts
@@ -3,7 +3,7 @@ import { AuthController } from './auth.controller';
 import { RefreshToken, User } from '@prisma/client';
 import { AuthService } from './auth.service';
 import { UsersService } from '../users/users.service';
-import { RefreshTokensService } from './refresh-tokens.service';
+import { RefreshTokensService } from '../refresh-tokens/refresh-tokens.service';
 import { ProfilesService } from '../profiles/profiles.service';
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
