+++ nestjs-BE/server/src/refresh-tokens/refresh-tokens.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { RefreshTokensService } from './refresh-tokens.service';
+import { JwtModule } from '@nestjs/jwt';
+import { PrismaModule } from '../prisma/prisma.module';
+
+@Module({
+  imports: [JwtModule, PrismaModule],
+  providers: [RefreshTokensService],
+  exports: [RefreshTokensService],
+})
+export class RefreshTokensModule {}
